Добавляет функцию сравнения рабочего дня и продолжительности встречи.
